requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
greenlet==2.0.2
idna==3.4
psycopg2-binary==2.9.5
python-telegram-bot==13.14
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
